server/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
pydantic>=2.11.7
python-dotenv>=1.1.1
pytest>=8.4.1
psutil>=5.9.0
fastapi-mcp>=0.4.0
boto3>=1.35.90
numpy>=1.26.3
aiofiles>=23.2.1
python-multipart>=0.0.6
openai>=1.55.3
mutagen>=1.47.0
moviepy==1.0.3
requests>=2.31.0
Pillow>=10.0.0